Jenkinsfile new

+++ Jenkinsfile
@@ -19,9 +19,14 @@ pipeline {
         }
         stage('Apache2 install') {
             steps {
-                sh 'apt-get remove -y apache2'
-                sh 'apt-get purge -y apache2'
-                sh 'apt-get install -y apache2'
+                script {
+                    if (sh ('find /etc/init.d/* apache2') ) {
+                        sh 'apt-get remove -y apache2'
+                        sh 'apt-get purge -y apache2'
+                    } else {
+                        sh 'apt-get install -y apache2'
+                    }
+                }
             }
         }
         stage('Start') {
